pyproject.toml
[project]
name = "briangaudino"
requires-python = ">= 3.10"
dynamic = ["version"]
dependencies = [
    "django-admin-ordering>=0.20.0",
    "django>=5.2",
]

[tool.uv]
dev-dependencies = [
    "djlint>=1.36.4",
]

[tool.djlint]
indent = 2
